Add files via upload
The scope of the e-commerce website project encompasses the development of a user-friendly online platform allowing customers to register, browse products, add items to their cart, securely process payments, and manage orders .An ecommerce website caters to the exchange of goods and services for electronic transaction of money as well as Cash-On-Delivery (COD). People can visit your website, choose the products they like, enter the required details, make payment (or choose to pay later), and purchase goods and services in a jiffy.